+++ src/integrations/ExpressFormsIntegration.php
@@ -14,13 +14,22 @@ class ExpressFormsIntegration {
   public function parse(string $integration): void {
     $this->integration = $integration;
     Event::on(\Solspace\ExpressForms\models\Form::class, \Solspace\ExpressForms\models\Form::EVENT_VALIDATE_FORM, function(\Solspace\ExpressForms\events\forms\FormValidateEvent $e){
-      if (!$e->getForm()->isValid()){
+      $form = $e->getForm();
+      if (!$form->isValid()){
         return;
+      }
+      $handle = $form->handle;
+      if (isset(OOPSpam::$plugin->settings->forms[$handle])){
+        $settings = OOPSpam::$plugin->settings->forms[$handle];
+        if ((isset($settings['disabled'])) && ($settings['disabled'])){
+          return;
+        }
+        OOPSpam::overrideSettings($settings);
       }
       $params = [
         'content' => []
       ];
-      foreach($e->getForm()->getFields() as $field){
+      foreach($form->getFields() as $field){
         switch(get_class($field)){
           case 'Solspace\ExpressForms\fields\Email':
             $params['email'] = $field->getValue();
@@ -37,7 +46,7 @@ class ExpressFormsIntegration {
         $params['contextual'] = true;
       }
       if (!OOPSpam::$plugin->antiSpam->checkSpam($params, $this->getName())){
-        $e->getForm()->markAsSpam();
+        $form->markAsSpam();
       }
     });
   }